migrate SchemaDefinition
